fix: fix apt failed message redirecting to Discord
If apt update fails, the installer redirects to our Discord. As this is
not a problem with our installer or a problem we can influence, and
often a matter of just a quick google search, we decided to remove the
redirect to our Discord.
This should incentivise the user to learn more about such problems and
the reason why it appears.
We will ofc still help with any problems people may have on our Discord
or in our GitHub Discussions.